// generated/orval/fetch/google-health-api/models/heartRateZoneHeartRateZoneType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Google Health API
 * The Google Health API lets you view and manage health and fitness metrics and measurement data.
 * OpenAPI spec version: v4
 */

/**
 * Required. The heart rate zone type.
 */
export type HeartRateZoneHeartRateZoneType = typeof HeartRateZoneHeartRateZoneType[keyof typeof HeartRateZoneHeartRateZoneType];


export const HeartRateZoneHeartRateZoneType = {
  HEART_RATE_ZONE_TYPE_UNSPECIFIED: 'HEART_RATE_ZONE_TYPE_UNSPECIFIED',
  LIGHT: 'LIGHT',
  MODERATE: 'MODERATE',
  VIGOROUS: 'VIGOROUS',
  PEAK: 'PEAK',
} as const;