// generated/orval/fetch/google-health-api/models/heartRateZone.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Google Health API
 * The Google Health API lets you view and manage health and fitness metrics and measurement data.
 * OpenAPI spec version: v4
 */
import type { HeartRateZoneHeartRateZoneType } from './heartRateZoneHeartRateZoneType';

/**
 * The heart rate zone.
 */
export interface HeartRateZone {
  /** Required. The heart rate zone type. */
  heartRateZoneType?: HeartRateZoneHeartRateZoneType;
  /** Required. Maximum heart rate for this zone in beats per minute. */
  maxBeatsPerMinute?: string;
  /** Required. Minimum heart rate for this zone in beats per minute. */
  minBeatsPerMinute?: string;
}